simba/pyproject.toml
[tool.poetry]
name = "simba-core"
version = "0.4.0"
description = "Knowledge Management System that connects to your RAG system"
authors = ["GitHamza0206 <zeroualihamza0206@gmail.com>"]
readme = "README.md"
license = "LICENCE.md"
packages = [
    {include = "simba"},
    {include = "simba_sdk"}
]
repository = "https://github.com/GitHamza0206/simba"
keywords = ["rag", "knowledge-management", "vector-store", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langchain = "*"
# langgraph = "^0.2.60"
python-dotenv = "^1.0.1"
PyYAML = "^6.0.1" # Added for YAML configuration parsing
faiss-cpu = "*"
# langchain-community = "^0.3.13"
# langchain-openai = "^0.2.14"
uvicorn = "^0.33.0"
fastapi = "^0.115.6"
markdown = "^3.7"
unstructured = {version = "^0.11.0"}
# langchain-"Placeholder" = "^0.1.4"
markitdown = "0.0.1a1"
# Temporarily exclude pdfplumber due to pycrypto incompatibility
# unstructured[pdf] would normally pull it in
litellm = "^1.56.6"
python-multipart = "^0.0.20"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pillow-heif = "^0.18.0"
pdf2image = "^1.17.0"
pdfminer-six = "^20240706"
# docling = "^2.15.1"
# langchain-docling = "^0.2.0"
ipykernel = "^6.29.5"
ollama = "^0.4.6"
# langchain-ollama = "^0.2.2"
rq = "^2.1.0"
pymupdf = "^1.24.0"
rank-bm25 = "^0.2.2"
# langchain-experimental = "^0.3.4"
sentence-transformers = "^3.2.0"
# langchain-huggingface = "^0.1.2"
pydantic-settings = "^2.7.1"
torch = "^2.4.0"
torchvision = ">=0.19.0"
torchaudio = "^2.4.0"
celery = "^5.4.0"
mistralai = "^1.5.1"
minio = "^7.2.0"
supabase = "^2.6.0"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
pgvector = "^0.4.1"
email-validator = "^2.2.0"
cryptography = ">=36.0.0"
pycryptodome = "*"
ag2 = "*"
# Ensure pycrypto is not installed, use pycryptodome instead
mcp = "^1.12.4"
[tool.poetry.extras]
no-pycrypto = ["pycryptodome"]


[tool.poetry.group.dev.dependencies]
nox = "^2025.2.9"
isort = "^6.0.0"
autoflake = "^2.3.1"
pre-commit = "^4.1.0"
flake8 = "^7.1.2"
black = "^25.1.0"
uv = "^0.6.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
mypy = "^1.15.0"
bandit = "^1.8.3"
safety = "^3.3.1"
sphinx = "^8.2.1"
sphinx-rtd-theme = "^3.0.2"
twine = "^6.1.0"
build = "^1.2.2.post1"
python-semantic-release = "^9.21.0"


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".git", "venv", ".venv", "env", ".env", "__pycache__", "dist", "build"]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
simba = "simba.cli:main"
